% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_active_x.R
\name{callActiveXClp}
\alias{callActiveXClp}
\title{phase x chromosome SNPs}
\usage{
callActiveXClp(
  outErrorRatePath,
  outXCallsPath,
  preparedDacPath,
  phasePath,
  max.error.rate = 0.25,
  min.cells.final = 4,
  calling.purity.min = 0.85,
  likelihood.min = 0.9
)
}
\arguments{
\item{outErrorRatePath}{path to write error rate table}

\item{outXCallsPath}{path to write active-X calls}

\item{preparedDacPath}{file produced by prepareDacClp()}

\item{phasePath}{file produced by phaseClp()

TODO: Steve: document the rest of the parameters}
}
\description{
phase x chromosome SNPs
}
