% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{default_X_contig_name}
\alias{default_X_contig_name}
\title{Default name of the X chromosome: "X"}
\format{
An object of class \code{character} of length 1.
}
\usage{
default_X_contig_name
}
\description{
Default name of the X chromosome: "X"
}
\keyword{datasets}
