% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_and_merge_dacs}
\alias{load_and_merge_dacs}
\title{Load one or more DACs, disambiguate the CBCs based on the file names, and merge them into a single data.table.}
\usage{
load_and_merge_dacs(files)
}
\arguments{
\item{files}{one or more files with extension .dac.txt.gz, produced by GatherDigitalAlleleCounts.  The colnames should be
chr    pos    gene    cell    ref_base    alt_base    rA    rC    rG    rT    rN    A    C    G    T    N    num_umi    umi_mean_purity    read_pval    read_ratio    read_ci_low    read_ci_high    umi_pval    umi_ratio    umi_ci_low    umi_ci_high}
}
\description{
Load one or more DACs, disambiguate the CBCs based on the file names, and merge them into a single data.table.
}
