% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phaseClp}
\alias{phaseClp}
\title{phase x chromosome SNPs}
\usage{
phaseClp(
  outPhasePath,
  preparedDacPath,
  min.cells = 2,
  max.xi.proportion = 0.1,
  max.gen.units = 1000,
  X_contig_name = default_X_contig_name
)
}
\arguments{
\item{outPhasePath}{path to write x SNP phase table}

\item{preparedDacPath}{file produced by prepareDacClp()

TODO: Steve: document the rest of the parameters}
}
\description{
phase x chromosome SNPs
}
