% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dac.R
\name{prepareDacClp}
\alias{prepareDacClp}
\title{Join DACs and VCF, annotate with GTF, and prepare for phasing}
\usage{
prepareDacClp(
  outDacPath,
  dacPaths,
  annotationsPath,
  vcfPath,
  intronicOnly = TRUE,
  X_contig_name = default_X_contig_name
)
}
\arguments{
\item{outDacPath}{path to write output DAC}

\item{dacPaths}{paths to DAC files, one for each reaction.  Filenames are used to disambiguate CBCs}

\item{annotationsPath}{reduced GTF with columns chr,start,end,gene,annotation}

\item{vcfPath}{X-only VCF created with prepareVcfClp(), with columns pos    ref,alt,ps,gt,ad,dp,gq,pl (and if gnomAD:    gnomad.ac,gnomad.an,gnomad.af,gnomad.maf)}

\item{intronicOnly}{logical, keep only intronic reads?}

\item{X_contig_name}{name of the X contig}
}
\description{
Join DACs and VCF, annotate with GTF, and prepare for phasing
}
