% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gtf.R
\name{prepareGtfClp}
\alias{prepareGtfClp}
\title{Prepare GTF gene annotations for xipher.}
\usage{
prepareGtfClp(
  outAnnotationsPath,
  inReducedGtfPath,
  X_contig_name = default_X_contig_name
)
}
\arguments{
\item{outAnnotationsPath}{path to write gene annotations in format used by prepareDacClp()}

\item{inReducedGtfPath}{path to reduced GTF file produced by ReduceGtf (see https://github.com/broadinstitute/Drop-seq/).
Note that you can speed ReduceGtf and make it's output smaller by providing a sequence dictionary
that contains only the X contig.}

\item{X_contig_name}{name of the X contig. Default: X}
}
\description{
Prepare GTF gene annotations for xipher.
}
