% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_table_helper}
\alias{write_table_helper}
\title{Helper function for writing a possibly gzipped data.table/data.frame}
\usage{
write_table_helper(
  outPath,
  x,
  col.names = TRUE,
  row.names = FALSE,
  quote = FALSE,
  sep = "\\t",
  ...
)
}
\arguments{
\item{outPath}{path to output file}

\item{x}{data.table/data.frame to write}

\item{col.names}{whether to write column names}

\item{row.names}{whether to write row names}

\item{quote}{whether to quote strings}

\item{sep}{column separator}

\item{...}{additional arguments to write.table}
}
\description{
Helper function for writing a possibly gzipped data.table/data.frame
}
