% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vcf.R
\name{prepareVcfClp}
\alias{prepareVcfClp}
\title{load donor vcf, prepare for x inference, including gnomAD filtering if provided.}
\usage{
prepareVcfClp(
  outVcfPath,
  inVcfPath,
  isMouse = FALSE,
  gnomAdPath = NULL,
  X_contig_name = default_X_contig_name
)
}
\arguments{
\item{outVcfPath}{path to write output vcf}

\item{inVcfPath}{path to input vcf}

\item{isMouse}{logical, is the input vcf from a mouse?}

\item{gnomAdPath}{path to gnomAD vcf}

\item{X_contig_name}{name of the X contig in the reference genome. Default: X}
}
\description{
load donor vcf, prepare for x inference, including gnomAD filtering if provided.
}
